requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.2
fonttools==4.55.8
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
